pyproject.toml
[tool.poetry]
name = "knowthyself"
version = "0.1.0"
description = "Gather all of your personal data from around the web and learn something about yourself."
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.13"
django-allauth = {extras = ["socialaccount"], version = "^64.0.0"}
python-webpack-boilerplate = "^1.0.0"
django-widget-tweaks = "^1.4.12"
django = "^5.0.4"
django-environ = "^0.11.2"
ipython = "^8.27.0"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
django-q2 = "^1.7.2"
whitenoise = "^6.7.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
structlog = "^24.4.0"
django-structlog = "^8.1.0"
markdown = "^3.7"
sentry-sdk = {extras = ["django"], version = "^2.14.0"}
structlog-sentry = "^2.2.1"
gunicorn = "^23.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
redis = "^5.0.8"
django-anymail = {extras = ["mailgun"], version = "^12.0"}
posthog = "^5.3.0"
dj-stripe = "^2.9.0"
stripe = "^11.6.0"
django-ninja = "^1.3.0"
django-mjml = "^1.3"
pydantic-ai = "^0.3.4"
logfire = "^3.6.4"
psycopg = {extras = ["binary"], version = "^3.2.10"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
djlint = "^1.36.4"
ruff = "^0.12.0"
ty = "^0.0.1a11"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
    "DJ",   # django-specific rules
    "UP",   # pyupgrade
    "C90",  # mccabe complexity
]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "F401", "F403", "F405"]

[tool.ruff.format]
quote-style = "double"

[tool.djlint]
profile="django"
ignore = "H031,H006,H023,H021,H011,T002"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
